# frontend/public/dev/openapi.json
{
  "components": {
    "schemas": {
...
      "OrdersHTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/OrdersValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrdersOrderResponse": {
        "description": "Public representation of an order.",
        "properties": {
          "checkout_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkout Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "discount_amount": {
            "default": "0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Discount Amount",
            "type": "string"
          },
          "drop_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drop Id"
          },
          "final_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Price"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "original_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Price"
          },
          "payment_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Expires At"
          },
          "payment_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Id"
          },
          "price": {
            "title": "Price",
            "type": "integer"
          },
          "product_id": {
            "format": "uuid",
            "title": "Product Id",
            "type": "string"
          },
          "product_name": {
            "title": "Product Name",
            "type": "string"
          },
          "promocode_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promocode Id"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "reservation_id": {
            "format": "uuid",
            "title": "Reservation Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OrdersOrderStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "variant_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Color"
          },
          "variant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Id"
          },
          "variant_size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Size"
          },
          "variant_sku": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Sku"
          }
        },
        "required": [
          "id",
          "user_id",
          "product_id",
          "product_name",
          "price",
          "currency",
          "quantity",
          "status",
          "reservation_id",
          "payment_id",
          "created_at",
          "updated_at"
        ],
        "title": "OrderResponse",
        "type": "object"
      },
      "OrdersOrderStatus": {
        "description": "Lifecycle status of an order.",
        "enum": [
          "PENDING",
          "AWAITING_PAYMENT",
          "PAID",
          "CONFIRMED",
          "PAYMENT_FAILED",
          "CANCELLED"
        ],
        "title": "OrderStatus",
        "type": "string"
      },
...
      "OrdersValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "PaymentsCheckoutResponse": {
        "description": "Browser-safe data needed to continue a hosted payment.",
        "properties": {
          "attempt_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id"
          },
          "confirmation_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Url"
          },
          "payment_id": {
            "format": "uuid",
            "title": "Payment Id",
            "type": "string"
          },
          "preparation_status": {
            "default": "ready",
            "title": "Preparation Status",
            "type": "string"
          },
          "retry_after_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Seconds"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentsPaymentStatus"
          }
        },
        "required": [
          "payment_id",
          "status"
        ],
        "title": "CheckoutResponse",
        "type": "object"
      },
      "PaymentsCreatePaymentRequest": {
        "description": "Payload for creating a payment for an order.",
        "properties": {
          "amount": {
            "exclusiveMinimum": 0.0,
            "maximum": 10000000000.0,
            "title": "Amount",
            "type": "integer"
          },
          "currency": {
            "default": "RUB",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "order_id": {
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          },
          "provider": {
            "default": "mock",
            "maxLength": 64,
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "order_id",
          "user_id",
          "amount"
        ],
        "title": "CreatePaymentRequest",
        "type": "object"
      },
      "PaymentsHTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/PaymentsValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PaymentsPaymentListResponse": {
        "description": "Paginated payment listing.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PaymentsPaymentResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaymentListResponse",
        "type": "object"
      },
      "PaymentsPaymentResponse": {
        "description": "Public representation of a payment.",
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "integer"
          },
          "cancellation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Reason"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "current_attempt_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Attempt Id"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "external_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Status"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "order_id": {
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "provider_test": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Test"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentsPaymentStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_id",
          "user_id",
          "amount",
          "currency",
          "provider",
          "status",
          "external_id",
          "expires_at",
          "created_at",
          "updated_at"
        ],
        "title": "PaymentResponse",
        "type": "object"
      },
      "PaymentsPaymentStatus": {
        "description": "Lifecycle states of a payment attempt.",
        "enum": [
          "PENDING",
          "SUCCESS",
          "FAILED",
          "CANCELLED",
          "REFUNDED"
        ],
        "title": "PaymentStatus",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Same-origin public API contract for the FlashMarket platform.",
    "title": "FlashMarket Public API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/orders/{order_id}": {
      "get": {
        "description": "Return a single order by id.",
        "operationId": "get_order_api_v1_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersOrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersHTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get an order",
        "tags": [
          "orders"
        ],
        "x-flashmarket-access": "authenticated",
        "x-flashmarket-service": "orders"
      }
    },
...
    "/api/v1/payments": {
      "post": {
        "description": "Create a pending payment for an administrative mock workflow.",
        "operationId": "create_payment_api_v1_payments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentsCreatePaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsPaymentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsHTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a payment for an order",
        "tags": [
          "payments"
        ],
        "x-flashmarket-access": "admin",
        "x-flashmarket-service": "payments"
      }
    },
    "/api/v1/payments/orders/{order_id}": {
      "get": {
        "description": "Return the authoritative payment for an order.",
        "operationId": "get_order_payment_api_v1_payments_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsPaymentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsHTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get payment by order",
        "tags": [
          "payments"
        ],
        "x-flashmarket-access": "authenticated",
        "x-flashmarket-service": "payments"
      }
    },
    "/api/v1/payments/orders/{order_id}/checkout": {
      "post": {
        "description": "Create a provider payment from the authoritative PaymentRequested event.",
        "operationId": "start_checkout_api_v1_payments_orders__order_id__checkout_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsCheckoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsCheckoutResponse"
                }
              }
            },
            "description": "Provider result is being reconciled"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsHTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Start or resume hosted checkout",
        "tags": [
          "payments"
        ],
        "x-flashmarket-access": "authenticated",
        "x-flashmarket-service": "payments"
      }
    },
    "/api/v1/payments/users/{user_id}": {
      "get": {
        "description": "Return paginated payments for a user.",
        "operationId": "list_payments_api_v1_payments_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsPaymentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsHTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List user payments",
        "tags": [
          "payments"
        ],
        "x-flashmarket-access": "authenticated",
        "x-flashmarket-service": "payments"
      }
    },
...
    "/api/v1/payments/{payment_id}": {
      "get": {
        "description": "Return a single payment by id.",
        "operationId": "get_payment_api_v1_payments__payment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Payment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsPaymentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsHTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a payment",
        "tags": [
          "payments"
        ],
        "x-flashmarket-access": "authenticated",
        "x-flashmarket-service": "payments"
      }
    },
    "/api/v1/payments/{payment_id}/cancel": {
      "post": {
        "description": "Cancel a pending payment.",
        "operationId": "cancel_payment_api_v1_payments__payment_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Payment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsPaymentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsHTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Cancel payment",
        "tags": [
          "payments"
        ],
        "x-flashmarket-access": "authenticated",
        "x-flashmarket-service": "payments"
      }
    },
    "/api/v1/payments/{payment_id}/confirm": {
      "post": {
        "description": "Mark a pending payment as successful (admin/provider callback only).",
        "operationId": "confirm_payment_api_v1_payments__payment_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Payment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsPaymentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsHTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Confirm payment succeeded",
        "tags": [
          "payments"
        ],
        "x-flashmarket-access": "admin",
        "x-flashmarket-service": "payments"
      }
    },
    "/api/v1/payments/{payment_id}/fail": {
      "post": {
        "description": "Mark a pending payment as failed (admin/provider callback only).",
        "operationId": "fail_payment_api_v1_payments__payment_id__fail_post",
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Payment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsPaymentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsHTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Fail payment",
        "tags": [
          "payments"
        ],
        "x-flashmarket-access": "admin",
        "x-flashmarket-service": "payments"
      }
    },
...
  },
  "servers": [
    {
      "description": "Current FlashMarket domain",
      "url": "/"
    }
  ],
...
}